Refactor config and modules
Memory (RAM) usage improvements

+++ include/platform.h
@@ -15,7 +15,7 @@ namespace platform {
 	void loop();
 	bool isConfigured();
 	bool isConnected();
-	double getExchangeRate();
+	float getExchangeRate();
 }
 
 #endif
